+++ modules/sdk/gradle-plugins/CHANGELOG.markdown
@@ -1641,6 +1641,46 @@ to version 2.0.7.
 - [LPS-72656]: Update the [Liferay Source Formatter] dependency to version
 1.0.397.
 
+## 3.3.16 - 2017-05-23
+
+### Changed
+- [LPS-67352]: Update the [Liferay Gradle Plugins Source Formatter] dependency
+to version 2.0.8.
+- [LPS-67352]: Update the [Liferay Source Formatter] dependency to version
+1.0.398.
+- [LPS-72723]: Update the [Liferay Gradle Plugins JS Transpiler] dependency to
+version 2.3.13.
+
+## 3.3.17 - 2017-05-23
+
+### Changed
+- [LPS-67352]: Update the [Liferay Gradle Plugins Source Formatter] dependency
+to version 2.0.9.
+- [LPS-67352]: Update the [Liferay Source Formatter] dependency to version
+1.0.399.
+
+## 3.3.18 - 2017-05-23
+
+### Changed
+- [LPS-71722]: Update the [Liferay Gradle Plugins Service Builder] dependency
+to version 1.0.40.
+- [LPS-71722]: Update the [Liferay Portal Tools Service Builder] dependency to
+version 1.0.156.
+
+## 3.3.19 - 2017-05-23
+
+### Changed
+- [LPS-71722]: Update the [Liferay Gradle Plugins Service Builder] dependency
+to version 1.0.41.
+- [LPS-71722]: Update the [Liferay Portal Tools Service Builder] dependency to
+version 1.0.157.
+
+## 3.3.20 - 2017-05-24
+
+### Changed
+- [LPS-72750]: Update the [Liferay Gradle Plugins Soy] dependency to version
+3.1.3.
+
 [Gradle Bundle Plugin]: https://github.com/TomDmitriev/gradle-bundle-plugin
 [Liferay Alloy Taglib]: https://github.com/liferay/alloy-taglibs
 [Liferay Ant BND]: https://github.com/liferay/liferay-portal/tree/master/modules/sdk/ant-bnd
@@ -1790,4 +1830,6 @@ to version 2.0.7.
 [LPS-72514]: https://issues.liferay.com/browse/LPS-72514
 [LPS-72572]: https://issues.liferay.com/browse/LPS-72572
 [LPS-72656]: https://issues.liferay.com/browse/LPS-72656
+[LPS-72723]: https://issues.liferay.com/browse/LPS-72723
+[LPS-72750]: https://issues.liferay.com/browse/LPS-72750
 [LRDOCS-3023]: https://issues.liferay.com/browse/LRDOCS-3023